Moving enums into more logical files
