--- src/Types.tsx
@@ -1,26 +1,0 @@
-export enum Mode {
-  Normal = 'NORMAL',
-  Assay = 'ASSAY',
-  Add = 'ADD',
-  Subtract = 'SUBTRACT'
-}
-
-export enum OrganelleType {
-  Nucleus = 'NUCLEUS',
-  Cytoplasm = 'CYTOPLASM',
-  Golgi = 'GOLGI',
-  Gates = 'GATES',
-  Intercell = 'INTERCELL'
-}
-
-export enum SubstanceType {
-  Substance1 = 'Substance 1',
-  Substance2 = 'Substance 2',
-  Substance3 = 'Substance 3'
-}
-
-export enum View {
-  None = 'NONE',
-  Organism = 'ORGANISM',
-  Cell = 'CELL'
-}
